add tests, move to utils file
